# UI/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TF26 API",
    "description": "Turkish participation-banking campaigns and live pricing. This service authenticates callers and exposes what `banks/`, `index/` and `corpus/` already do; it holds no banking logic itself.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/banks/{bank}": {
      "get": {
        "tags": [
          "banks"
        ],
        "summary": "One Bank",
        "operationId": "one_bank",
        "parameters": [
          {
            "name": "bank",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "A bank key from GET /api/banks.",
              "title": "Bank"
            },
            "description": "A bank key from GET /api/banks."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/banks/{bank}/products": {
      "get": {
        "tags": [
          "banks"
        ],
        "summary": "Bank Products",
        "description": "A bank's own catalogue, as the bank publishes it.",
        "operationId": "bank_products",
        "parameters": [
          {
            "name": "bank",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "A bank key from GET /api/banks.",
              "title": "Bank"
            },
            "description": "A bank key from GET /api/banks."
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "finance | profit_share",
              "default": "finance",
              "title": "Category"
            },
            "description": "finance | profit_share"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductOut"
                  },
                  "title": "Response Bank Products"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/banks/{bank}/finance": {
      "get": {
        "tags": [
          "banks"
        ],
        "summary": "Bank Finance Quote",
        "description": "A live financing quote, including the full payment schedule.\n\nThe number is the bank's own. Nothing here recomputes an installment or\nfills a gap with a formula -- an endpoint that will not answer is reported\nas not answering.",
        "operationId": "bank_finance_quote",
        "parameters": [
          {
            "name": "bank",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "A bank key from GET /api/banks.",
              "title": "Bank"
            },
            "description": "A bank key from GET /api/banks."
          },
          {
            "name": "product",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "A product code from /products.",
              "title": "Product"
            },
            "description": "A product code from /products."
          },
          {
            "name": "amount",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "exclusiveMinimum": 0,
              "title": "Amount"
            }
          },
          {
            "name": "term",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 360,
              "exclusiveMinimum": 0,
              "description": "Months.",
              "title": "Term"
            },
            "description": "Months."
          },
          {
            "name": "monthly_profit_rate",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 100,
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Monthly Profit Rate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinanceQuoteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/banks/{bank}/profit-share": {
      "get": {
        "tags": [
          "banks"
        ],
        "summary": "Bank Profit Share Quote",
        "description": "A live participation-account quote.\n\nBanks disagree about whether a product takes a term in days or months, so\nthe response repeats the unit that actually answered rather than echoing\nthe one that was asked for.",
        "operationId": "bank_profit_share_quote",
        "parameters": [
          {
            "name": "bank",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "A bank key from GET /api/banks.",
              "title": "Bank"
            },
            "description": "A bank key from GET /api/banks."
          },
          {
            "name": "product",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product"
            }
          },
          {
            "name": "amount",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "exclusiveMinimum": 0,
              "title": "Amount"
            }
          },
          {
            "name": "term",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Term"
            }
          },
          {
            "name": "unit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(day|month)$",
              "default": "month",
              "title": "Unit"
            }
          },
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 3,
              "default": "TRY",
              "title": "Currency"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfitShareQuoteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/banks/{bank}/rates": {
      "get": {
        "tags": [
          "banks"
        ],
        "summary": "Bank Rates",
        "description": "A bank's FX and gold table, with the timestamp it was quoted at.\n\nCached for a few seconds, and only here. The board is a live page: it polls,\nand without this every tab refreshing would be its own request to the bank.\nTwo of these boards are page reads and one bank fingerprints TLS, so the\nbrowser polls freely and the banks see one request per TTL.",
        "operationId": "bank_rates",
        "parameters": [
          {
            "name": "bank",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "A bank key from GET /api/banks.",
              "title": "Bank"
            },
            "description": "A bank key from GET /api/banks."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RateOut"
                  },
                  "title": "Response Bank Rates"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/banks/{bank}/card": {
      "get": {
        "tags": [
          "banks"
        ],
        "summary": "Bank Card Quote",
        "description": "An instalment plan for a credit-card purchase.\n\nOnly two banks publish this (Kuveyt Türk, Vakıf). Kuveyt Türk's own card\ncatalogue has a duplicate code -- `BP` names two different cards -- so a\ncaller quoting `BP` gets whichever the catalogue lists first; this is a\nknown gap in the bank's own data, not something resolved here.",
        "operationId": "bank_card_quote",
        "parameters": [
          {
            "name": "bank",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "A bank key from GET /api/banks.",
              "title": "Bank"
            },
            "description": "A bank key from GET /api/banks."
          },
          {
            "name": "card",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "A card code from GET /{bank}/products?category=card.",
              "title": "Card"
            },
            "description": "A card code from GET /{bank}/products?category=card."
          },
          {
            "name": "amount",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "exclusiveMinimum": 0,
              "title": "Amount"
            }
          },
          {
            "name": "installments",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Installments"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardInstallmentQuoteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/banks/{bank}/miles": {
      "get": {
        "tags": [
          "banks"
        ],
        "summary": "Bank Mile Rates",
        "description": "Miles earned per lira, by card, tier and spending category.\n\nKuveyt Türk is the only publisher, and its table is 567 rows -- every\ncombination of card, membership tier and category. Filtering by card or\ncategory happens in the frontend rather than as query parameters here,\nwhich would need their own \"list the valid values\" endpoints to be usable.",
        "operationId": "bank_mile_rates",
        "parameters": [
          {
            "name": "bank",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "A bank key from GET /api/banks.",
              "title": "Bank"
            },
            "description": "A bank key from GET /api/banks."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MileRateOut"
                  },
                  "title": "Response Bank Mile Rates"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BankOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The key every other endpoint takes."
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "publishes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Publishes"
          },
          "finance_input_capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Finance Input Capabilities"
          },
          "maintenance": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Maintenance",
            "description": "Capabilities the last health check found failing."
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "name",
          "display_name",
          "publishes"
        ],
        "title": "BankOut",
        "description": "One bank, what it publishes, and what is currently down.\n\n`publishes` and `maintenance` are different absences and are reported\nseparately: a bank that does not sell car finance and a bank whose car\nfinance endpoint is broken this morning must never look alike in the UI."
      },
...
      "FinanceQuoteOut": {
        "properties": {
          "bank": {
            "type": "string",
            "title": "Bank"
          },
          "product": {
            "$ref": "#/components/schemas/ProductOut"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "term": {
            "type": "integer",
            "title": "Term"
          },
          "installment": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Installment",
            "description": "Monthly payment, or null where the bank publishes only a rate."
          },
          "total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          },
          "profit_rate": {
            "type": "number",
            "title": "Profit Rate"
          },
          "annual_cost_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Annual Cost Rate"
          },
          "fees": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Fees"
          },
          "variant": {
            "type": "string",
            "title": "Variant",
            "description": "e.g. \"sigortali\" / \"sigortasiz\". Empty where the bank does not split.",
            "default": ""
          },
          "general": {
            "type": "boolean",
            "title": "General",
            "default": false
          },
          "derived": {
            "type": "boolean",
            "title": "Derived",
            "default": false
          },
          "schedule": {
            "items": {
              "$ref": "#/components/schemas/PaymentRowOut"
            },
            "type": "array",
            "title": "Schedule"
          }
        },
        "type": "object",
        "required": [
          "bank",
          "product",
          "amount",
          "term",
          "profit_rate"
        ],
        "title": "FinanceQuoteOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProductOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "min_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Amount"
          },
          "max_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Amount"
          },
          "min_term": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Term"
          },
          "max_term": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Term"
          },
          "currencies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Currencies"
          },
          "rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "category"
        ],
        "title": "ProductOut"
      },
...
      "ProfitShareQuoteOut": {
        "properties": {
          "bank": {
            "type": "string",
            "title": "Bank"
          },
          "product": {
            "$ref": "#/components/schemas/ProductOut"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "term": {
            "type": "integer",
            "title": "Term"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "term_unit": {
            "type": "string",
            "title": "Term Unit"
          },
          "ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ratio"
          },
          "gross_profit": {
            "type": "number",
            "title": "Gross Profit"
          },
          "net_profit": {
            "type": "number",
            "title": "Net Profit"
          },
          "gross_annual_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gross Annual Rate"
          },
          "net_annual_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Annual Rate"
          },
          "variant": {
            "type": "string",
            "title": "Variant",
            "default": ""
          },
          "general": {
            "type": "boolean",
            "title": "General",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "bank",
          "product",
          "amount",
          "term",
          "currency",
          "term_unit",
          "gross_profit",
          "net_profit"
        ],
        "title": "ProfitShareQuoteOut"
      },
...
      "RateOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "buy": {
            "type": "number",
            "title": "Buy"
          },
          "sell": {
            "type": "number",
            "title": "Sell"
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "default": "1"
          },
          "quote_currency": {
            "type": "string",
            "title": "Quote Currency",
            "description": "Currency the bank used to quote one unit. The main FX board compares TRY quotes only.",
            "default": "TRY"
          },
          "as_of": {
            "type": "string",
            "title": "As Of",
            "default": ""
          },
          "derived": {
            "type": "boolean",
            "title": "Derived",
            "description": "True when the pair was worked out from the bank's own converter rather than read off a published feed. The UI must label these: the numbers are the bank's, the inversion is ours.",
            "default": false
          },
          "canonical": {
            "type": "string",
            "title": "Canonical",
            "description": "The shared symbol for this instrument -- XAU for gold whether the bank calls it XAU or 'ALT (gr)'. Cross-bank grouping keys on (canonical, unit); without it the frontend would have to duplicate the alias table and drift when a bank renames a code.",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "buy",
          "sell"
        ],
        "title": "RateOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}